Add more docker install steps

+++ linux/packages.sh
@@ -161,22 +161,30 @@ sudo apt-get -y install mysql-workbench-community
 # ---------------------------------------------------
 # Install Docker
 # ---------------------------------------------------
-sudo apt-get install -y \
-    apt-transport-https \
-    ca-certificates \
-    curl \
-    gnupg-agent \
-    software-properties-common
+if cmd_exists "docker"; then
+    skip "docker"
+else
+  sudo apt-get install -y \
+      apt-transport-https \
+      ca-certificates \
+      curl \
+      gnupg-agent \
+      software-properties-common
 
-curl -fsSL https://download.docker.com/linux/ubuntu/gpg | sudo apt-key add -
+  curl -fsSL https://download.docker.com/linux/ubuntu/gpg | sudo apt-key add -
 
-sudo add-apt-repository \
-   "deb [arch=amd64] https://download.docker.com/linux/ubuntu \
-   $(lsb_release -cs) \
-   stable"
+  sudo add-apt-repository \
+     "deb [arch=amd64] https://download.docker.com/linux/ubuntu \
+     $(lsb_release -cs) \
+     stable"
 
-sudo apt-get update
-sudo apt-get install -y docker-ce docker-ce-cli containerd.io
+  sudo apt-get update
+  sudo apt-get install -y docker-ce docker-ce-cli containerd.io
+
+  sudo groupadd docker
+  sudo usermod -aG docker $USER
+  newgrp docker
+fi
 # ---------------------------------------------------
